backend/requirement.txt
Flask
flask-marshmallow
Flask-Sqlalchemy
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
flask-cors
python-dotenv
werkzeug.security
psycopg2.extras
Flask-JWT-Extended==4.3.1